finished blocktags and loottables
Notes: The blocktags for logs and wood do not match exactly, this is because its a lot easyer in code this way. Modded blocks have 2 more tags, being natural logs (because some modded blocks WILL be natural logs.. and really when you think about it... a bear also scratches logs and turns them into stripped logs irl? so it should still be natural right?) and then also the tag for snaps goat horn is added. because..... well... its such a niche feature, and normal logs do snap it? so why not stripped logs? its not like the bark is harder then the wood itself.... so yeh.